Conflicts:
	weasis-base/weasis-base-ui/pom.xml
	weasis-core/weasis-core-api/src/main/java/org/weasis/core/api/media/MimeInspector.java
	weasis-core/weasis-core-api/src/main/java/org/weasis/core/api/media/data/SeriesThumbnail.java
	weasis-core/weasis-core-ui/pom.xml
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/DefaultMimeAppFactory.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/FileModel.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/SeriesViewerEvent.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/ViewerPluginBuilder.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/CrosshairListener.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/DefaultView2d.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/ExportImage.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/ImageTransferHandler.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/ImageViewerEventManager.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/ImageViewerPlugin.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/MeasureToolBar.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/MouseActions.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/ViewerToolBar.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/ZoomWin.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/dockable/MeasureTool.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/internal/Activator.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/model/layer/AbstractInfoLayer.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/GeneralSetting.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/JLocaleFormat.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/JLocaleLanguage.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/LabelsPrefView.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/PreferenceDialog.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/ScreenPrefView.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/StatisticsPrefView.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/pref/ViewSetting.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/util/ImagePrint.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/util/ToolBarContainer.java
	weasis-dicom/pom.xml
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/CheckTreeModel.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/DicomDirImport.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/DicomExplorer.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/DicomModel.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/DicomSorter.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/DicomZipImport.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/LocalExport.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/LocalImport.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/internal/Activator.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/pref/node/AbstractDicomNode.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/pref/node/DefaultDicomNode.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/pref/node/DicomQidoNode.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/pref/node/DicomWebNode.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/pref/node/DicomWebNodeDialog.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/print/DicomPrintDialog.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/wado/DicomManager.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/wado/DownloadManager.java
	weasis-dicom/weasis-dicom-explorer/src/main/java/org/weasis/dicom/explorer/wado/LoadSeries.java
	weasis-distributions/etc/config/config.properties
	weasis-distributions/pom.xml
	weasis-distributions/src/main/resources-portable/viewer-linux.sh
	weasis-launcher/conf/config.properties
	weasis-launcher/pom.xml
	weasis-launcher/src/main/java/org/weasis/launcher/LookAndFeels.java
	weasis-launcher/src/main/java/org/weasis/launcher/SwingResources.java
	weasis-launcher/src/main/java/org/weasis/launcher/WeasisLauncher.java
	weasis-launcher/src/main/java/org/weasis/launcher/WeasisLoader.java
	weasis-launcher/src/main/java/org/weasis/launcher/WebstartLauncher.java
	weasis-parent/pom.xml

+++ weasis-core/weasis-core-api/src/main/java/org/weasis/core/api/gui/task/SeriesProgressMonitor.java
@@ -42,7 +42,7 @@ public class SeriesProgressMonitor extends FilterInputStream {
         return loader == null || loader.isStopped();
     }
 
-    protected void updateSeriesProgression(double addSize) {
+    protected void updateSeriesProgression(long addSize) {
         series.setFileSize(series.getFileSize() + addSize);
         Thumbnail thumb = (Thumbnail) series.getTagValue(TagW.Thumbnail);
         if (thumb != null) {
@@ -62,7 +62,7 @@ public class SeriesProgressMonitor extends FilterInputStream {
         int c = in.read();
         if (c >= 0) {
             nread++;
-            updateSeriesProgression(1.0);
+            updateSeriesProgression(1);
         }
 
         return c;
